+++ zipflinger/src/com/android/zipflinger/FileSource.java
@@ -58,7 +58,7 @@ public class FileSource extends Source {
     }
 
     @Override
-    int writeTo(@NonNull ZipWriter writer) throws IOException {
+    long writeTo(@NonNull ZipWriter writer) throws IOException {
         return writer.write(buffer);
     }
 }
